This commit add the following functionality.
It allows for a tag-flow style release builds.

User may provide a regex, similar to the current pattern used in filters as a new parameter.

If the pipeline is triggered on a tagged commit the env var CIRCLE_TAG will be set.

If this var is set we will check that the current tag matches the reference tag, that the reference tag is a valid regex, that the previous tag matches the reference tag and the head tag per the regex. If all these align, we will set base to the previously tagged commit.

This is designed for teams using a trunk based flow on a monorepo. Where staging builds are deployed off merges to master and release builds are deployed via tagging.